requirements.txt
astroid==1.4.8
Babel==2.3.4
certifi==2017.1.23
cffi==1.9.1
click==6.7
cloudant==2.3.1
colorama==0.3.7
cryptography==1.7.1
cycler==0.10.0
debtcollector==1.11.0
decorator==4.0.10
entrypoints==0.2.2
Flask==0.12
gnupg==2.2.0
graphviz==0.5.1
idna==2.2
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
iso8601==0.1.11
isort==4.2.5
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
keystoneauth1==2.18.0
lazy-object-proxy==1.2.2
MarkupSafe==0.23
matplotlib==1.5.3
mccabe==0.5.2
mistune==0.7.3
monotonic==1.2
msgpack-python==0.4.8
nbconvert==4.2.0
nbformat==4.1.0
netaddr==0.7.19
netifaces==0.10.5
nltk==3.2.1
notebook==4.2.3
numpy==1.11.2
oslo.config==3.22.0
oslo.i18n==3.12.0
oslo.serialization==2.16.0
oslo.utils==3.22.0
pandas==0.19.1
pbr==1.10.0
pickleshare==0.7.4
positional==1.1.1
prompt-toolkit==1.0.9
psutil==5.0.1
pyasn1==0.1.9
pycparser==2.17
pyDes==2.0.1
Pygments==2.1.3
pylint==1.6.4
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-keystoneclient==3.9.0
python-swiftclient==3.2.0
pytz==2016.7
pyzmq==16.0.1
qtconsole==4.2.1
requests==2.12.5
rfc3986==0.4.1
scikit-learn==0.18.1
simplegeneric==0.8.1
six==1.10.0
stevedore==1.20.0
tornado==4.4.2
traitlets==4.3.1
urllib3==1.20
wcwidth==0.1.7
Werkzeug==0.11.15
widgetsnbextension==1.2.6
win-unicode-console==0.5
wrapt==1.10.8
